Issue 9 (#11)
* Stopped clearing the search string

After each batch file execution, the previous search was deleted.  This fix corrects that behavior.

Was not able to conveniently create a unit test without creating two additional, small methods in the primary codebase.

* search terms and selected position retained

After running a batch file, the search terms are kept and the highlight is in the same place as before.

Some tests are included, but not the main loop. Need to figure out how to unit test a "while True:" loop.

* Fixed a few PEP gripes

Resolves: #9